backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.32
alembic==1.13.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
